// frontend/src/lib/api/generated/model/workerRead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Construction ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { PayBasis } from './payBasis.ts';
import type { PayItemRead } from './payItemRead.ts';

export interface WorkerRead {
  /**
     * @minLength 1
     * @maxLength 120
     */
  full_name: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  trade: string;
  national_id?: string | null;
  phone?: string | null;
  pay_basis?: PayBasis;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  rate: string;
  id: string;
  is_active: boolean;
  pay_items?: PayItemRead[];
  created_at: string;
}